backend/src/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.8.0
channels==4.3.1
channels-redis==4.3.0
psycopg2-binary==2.9.10
redis==6.4.0
djangorestframework-simplejwt==5.5.1
django-axes==8.0.0
django-csp==4.0
celery==5.5.3
requests==2.32.5
gunicorn==23.0.0
uvicorn[standard]==0.34.0
daphne==4.2.1

